feat: progressInterval clamped from 0.1 to 5

+++ src/components/AmazonIvsViewNativeComponent.ts
@@ -36,7 +36,7 @@ export interface NativeProps extends ViewProps {
   maxBitrate?: Int32;
   initialBufferDuration?: Double;
   pipEnabled?: boolean;
-  progressInterval?: Int32;
+  progressInterval?: Double;
   playInBackground?: boolean;
   notificationTitle?: string;
   notificationText?: string;
